requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.1.1
django-cors-headers==3.2.0
django-webpack-loader==0.6.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
google-search-results==1.7.0
idna==2.8
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
urllib3==1.25.7
